Move customized_vue.js from vendor directory to shared directory
`app/javascript/vendor/customized_vue.js` never really made too much
sense, since it's not like the contents of that file were actually a
vendor library; it was just a tweaking of a vendor library. The
relatively recently added `app/javascript/shared/` library makes sense,
I think, since indeed the customized_vue.js export of `Vue` + our
modifications is indeed "shared" between the multiple Vue apps in this
repo.
