+++ app/javascript/packs/groceries_initializer.js
@@ -1,4 +1,4 @@
-import { renderApp } from 'vendor/customized_vue';
+import { renderApp } from 'shared/customized_vue';
 import Groceries from 'groceries/groceries.vue';
 import { groceryVuexStoreFactory } from 'groceries/store';
