+++ app/javascript/packs/home_app.js
@@ -1,4 +1,4 @@
-import { renderApp } from 'vendor/customized_vue';
+import { renderApp } from 'shared/customized_vue';
 import Home from 'home/home.vue';
 
 renderApp(Home);
