+++ app/javascript/shared/customized_vue.js
